+++ inst/tinytest/test-dvnames.R
@@ -28,3 +28,11 @@ m2 <- lm(x ~ y, data = d)
 dvnout <- dvnames(list(m1, m2, 1))
 nondvn <- list("y" = m1, "x" = m2, "Model" = 1)
 expect_identical(dvnout, nondvn)
+
+# dvnames transformation
+m1 <- lm(mpg ~ cyl + hp, data = mtcars)
+m2 <- lm(log(mpg+1) ~ cyl + hp, data = mtcars)
+a <- modelsummary(dvnames(list(m1, m2)), output = "data.frame")
+b <- modelsummary(dvnames(list(m1, m2), strip = TRUE), output = "data.frame")
+expect_true("log(mpg + 1)" %in% colnames(a))
+expect_false("log(mpg + 1)" %in% colnames(b))